[REFAC] 주문 승인 비동기 처리 로직 수정
